+++ tests/tools.ts
@@ -1,4 +1,8 @@
-/** 
+import { find, isNil, isEmpty } from 'ramda'
+
+import { Event } from '../src'
+
+/**
  * Function to force test to wait extra time to ensure, that the socket connect it properly OFF
  *
  * */
@@ -8,4 +12,12 @@ export async function wait(ms: number): Promise<string> {
       resolve('')
     }, ms)
   })
+}
+
+export function getEventByType(events: Event[], type: string): Event | undefined {
+  if (isNil(events) || isEmpty(events) || !type) return undefined
+
+  const finderFn = (el: Event): boolean => el.type === type
+
+  return find(finderFn, events)
 }